requirements.txt
Flask==3.1.2
joblib==1.5.1
pandas==2.3.3
Pillow==11.3.0
python-dotenv==1.1.1
Requests==2.32.5
torch==2.8.0
transformers==4.56.2
Werkzeug==3.1.3
scikit-learn
